refactor: selectedTodoProvider 분리
